+++ tests/StateDocumentTest.php
@@ -14,6 +14,7 @@ namespace Xabbuh\XApi\Model\Tests;
 use Xabbuh\XApi\Model\Activity;
 use Xabbuh\XApi\Model\Agent;
 use Xabbuh\XApi\Model\DocumentData;
+use Xabbuh\XApi\Model\InverseFunctionalIdentifier;
 use Xabbuh\XApi\Model\State;
 use Xabbuh\XApi\Model\StateDocument;
 
@@ -24,7 +25,7 @@ class StateDocumentTest extends AbstractDocumentTest
 {
     protected function createDocument(DocumentData $data)
     {
-        $agent = new Agent('mailto:alice@example.com');
+        $agent = new Agent(InverseFunctionalIdentifier::withMbox('mailto:alice@example.com'));
         $activity = new Activity('activity-id');
 
         return new StateDocument(new State($activity, $agent, 'state-id'), $data);